Updated discordgo to 0.23.0. Permissions changed to int64

+++ utils/utils.go
@@ -21,7 +21,7 @@ func HasPrefixes(str string, prefixes []string, caseSensitive bool) (bool, strin
 }
 
 //MemberHasPermission checks if guild member has a permission to do something on a server.
-func MemberHasPermission(s *discordgo.Session, guildID string, userID string, permission int) (bool, error) {
+func MemberHasPermission(s *discordgo.Session, guildID string, userID string, permission int64) (bool, error) {
 	member, err := s.State.Member(guildID, userID)
 	if err != nil {
 		if member, err = s.GuildMember(guildID, userID); err != nil {
